# stats: faster versions of SymbolTableImpl::lessThan and StatName::StartsWith