--- a/source/common/stats/symbol_table_impl.h
+++ b/source/common/stats/symbol_table_impl.h
@@ -73,6 +73,8 @@ public:
    */
   class Encoding {
   public:
+    class TokenIter;
+
     /**
      * Before destructing SymbolEncoding, you must call moveToMemBlock. This
      * transfers ownership, and in particular, the responsibility to call