@jmarantz 2022-01-28
Agreed - I think @ggreenway might have the most context - I think the Prometheus formatter he wrote might be the first beneficiary of this.